requirements.txt
google-generativeai
python-dotenv
Pillow
tqdm
pandas